fixing tabs and adding vendor

+++ scripts/test.sh
@@ -7,30 +7,31 @@ echo "Testing .go files"
 cd src/handlers/
 
 for folder in */;
-  do
-  (cd $folder
+	do
+		if [ $folder == "vendor/" ] ; then
+			continue;
+		fi
+		
+		(cd $folder
+			for innerFile in *;
+				do
+					if [ $innerFile == *".out" ] ; then
+						rm $innerFile
+					fi
 
-    for innerFile in *;
-    do
-      if [ $innerFile == *".out" ] ; then
-        rm $innerFile
-      fi
+					if [ $innerFile == *".html" ] ; then
+						rm $innerFile
+					fi
 
-      if [ $innerFile == *".html" ] ; then
-        rm $innerFile
-      fi
-
-      if [ $innerFile == *".pdf" ] ; then
-        rm $innerFile
-      fi
-    
-    done
+					if [ $innerFile == *".pdf" ] ; then
+						rm $innerFile
+					fi
+				done
       
-    go test -coverprofile=cover.out
-    go tool cover -html=cover.out -o cover.html
-    wkhtmltopdf cover.html cover.pdf
-
+      go test -coverprofile=cover.out
+      go tool cover -html=cover.out -o cover.html
+      wkhtmltopdf cover.html cover.pdf
     )
-  done
+	done
 
 echo "Done."
